+++ rector.php
@@ -5,9 +5,15 @@ declare(strict_types=1);
 use Rector\Config\RectorConfig;
 
 return RectorConfig::configure()
-    ->withPhpSets()
     ->withPaths([
         __DIR__ . '/example',
         __DIR__ . '/src',
         __DIR__ . '/tests',
-    ]);
+    ])
+    ->withPreparedSets(
+        deadCode: true,
+        codeQuality: true,
+        typeDeclarations: true,
+        earlyReturn: true,
+    )
+    ->withPhpSets();