+++ src/Exception/InvalidPaymentRequestException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mds\Natcash\Exception;
+
+/**
+ * Thrown when the provided payment request is invalid.
+ */
+class InvalidPaymentRequestException extends NatcashException {}